Profiling phase of sensitivity Test
The setup script was changed to enable all components of the model and make them play a role in the simulation.
A profiling of the code and identification of the bottleneck in the code will be possible that way.
The color scheme of the result graphs was also changed to make easier the distinction between the different figures.