Modify high slope value of decision function
1. Modification of high slope value: Change the the high slope value in
the decision function to 25% instead of 10% to allow more heterogeneity
in the decision of next position for grazers. Since the maximum size
the environement is limited to about 1 km of height and width, the
slope need to be considered at a more micro level compare to the more
macro level insinuated with the 10% slope value specified as critical
value for grazing preference of cattle in the litterature.
The 25% critical value is determined based on the fact it is just under
the shadow zone value of 15°. This allow to keep critical value for
high slopes consistent accross different modules of the model.

2. Modification vegetation figures: The colors of some vegetation
figure was modify to allow a better representation and be easier
to differenciate the differente components represented. The figure
of vegetation time series was modified to display the time axis (x)
as integer only.